pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-memory"
version = "0.1.0"
description = "A lightweight MCP memory server for AI agents. Markdown-first, zero external dependencies."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "OpenClaw Contributors" }]
keywords = ["mcp", "memory", "agent", "markdown", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp>=1.0.0",
    "sqlite-vec>=0.1.6",
    "python-frontmatter>=1.1.0",
    "watchfiles>=1.0.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
ollama = ["ollama>=0.4.0"]
local = ["sentence-transformers>=3.0.0"]
all = [
    "openclaw-memory[openai]",
    "openclaw-memory[ollama]",
    "openclaw-memory[local]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
openclaw-memory = "openclaw_memory.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_memory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
